Feature agregar categorias

+++ database/migrations/2025_03_19_202441_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name');
             $table->text('description');
             $table->string('num_reference');
-            $table->float('stock');
+            $table->integer('stock');
             $table->float('price');
             $table->string('image_url')->nullable(); // Campo para almacenar la URL de la imagen
